Update forTranslators.xml
Edited for perfect English. I'm not certain of the formatting though. Does any word with an ' need to be accompanied with a \? Like "can't or don't"? I saw a couple of times in there that they were like "can\'t" so I edited everything with an ' in it to be like that just to be sure. If I'm wrong about that, then all that needs done is to remove all of the \ in those words.